moved shortcuts hook
