+++ apps/web/src/components/navigation/command-palette.tsx
@@ -5,7 +5,7 @@ import { Dialog, DialogContent, DialogDescription, DialogHeader, DialogTitle } f
 import { Kbd } from '@/components/ui/kbd';
 import { useDialogContext } from '@/context/dialog-context';
 import type { Action } from '@/lib/actions';
-import { useShortcuts } from '@/lib/shortcuts';
+import { useShortcuts } from '@/hooks/use-shortcuts';
 import type { ReactNode } from 'react';
 
 type CommandPaletteProps = { actions: Action[] };
