+++ apps/web/src/components/ui/sidebar.tsx
@@ -14,7 +14,7 @@ import { Sheet, SheetContent, SheetDescription, SheetHeader, SheetTitle } from '
 import { Skeleton } from '@/components/ui/skeleton';
 import { Tooltip, TooltipContent, TooltipTrigger } from '@/components/ui/tooltip';
 import { useIsMobile } from '@/hooks/ui/use-mobile';
-import { useShortcuts } from '@/lib/shortcuts';
+import { useShortcuts } from '@/hooks/use-shortcuts';
 import { cn } from '@/lib/utils';
 
 const SIDEBAR_COOKIE_NAME = 'sidebar_state';
