+++ apps/web/src/hooks/use-shortcuts.ts
